Normalize discovered URIs before comparing them
python-openid2's urinorm is stricter than that in the old Python-2-only
version: in particular, it always normalizes the reserved character '+'
to '%2B'.  The discovery process normalizes URIs, so
GenericConsumer._verifyDiscoverySingle could end up comparing normalized
and unnormalized identifiers.

In particular, this broke with the Ubuntu single-sign-on system
(login.ubuntu.com / login.launchpad.net) which includes a '+' character
in the path part of its identifiers.

It makes sense to normalize both expected and discovered identifiers
before comparing them, so do that according to the rules in OpenID
Authentication 2.0 section 7.2.